+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Note - after update to 3.15.x need remove the accessory frome Home app and add it again.
 
+## [3.16.0] - (14.01.2022)
+### Added
+- ability to use inputs and surrounds mode with automations, shortcuts in HomeKit app [#131](https://github.com/grzegorz914/homebridge-denon-tv/issues/131)
+- ability to choice type of inputs or surrounds in automations (button, switch, motion sensor, ocupancy sensor)
+
+### Changs
+- code cleanup
+- removed all inputs and favorites function from buttons, now available in input section
+- update config.schema
+
+### Fixed
+- services calculation count
+- start input automation or scenes [#129](https://github.com/grzegorz914/homebridge-denon-tv/issues/129)
+
 ## [3.15.10] - (03.01.2022)
 ### Added
 -  ability to disable log device info by every connections device to the network (Advanced Section)